+++ Mods/SOD.LethalActionReborn/ModFolderContent/CHANGELOG.md
@@ -1,4 +1,9 @@
 # CHANGELOG
+**1.0.4**
+- Murderer and victim is now properly replaced if the person killed is one of them.
+- Added configuration option to define how many hits are needed after the citizen is KO'd to kill them.
+- Added configuration option to exclude certain weapon types from being able to kill citizens.
+
 **1.0.3**
 - Citizens will now be knocked out first, before you kill them.
 - Bugfix: Save/Load revived dead citizens.